+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/revisitor/logo_lang/revisitor-comp/test/impl/operation/impl/GreaterOpImpl.java
@@ -1,5 +1,6 @@
 package test.impl.operation.impl;
 
+import java.lang.Double;
 import kmLogo.Expression;
 import kmLogo.Greater;
 import kmLogo.Turtle;
@@ -59,10 +60,10 @@ public class GreaterOpImpl extends BinaryExpOpImpl implements GreaterOp {
   public double eval(Turtle turtle) {
     double result;
     if((rev.$((Expression)this.obj.getLhs()).eval(((Turtle) (turtle)))) > (rev.$((Expression)this.obj.getRhs()).eval(((Turtle) (turtle))))) {
-      result = 1.0;
+      result = ((Double) 1.0);
     }
     else {
-      result = 0.0;
+      result = ((Double) 0.0);
     }
     return result;
   }
